Explicit the cast, to calm down warnings about precision lost.
As far as I know, that is only guaranteed on Unix. On other platforms,
time_t may not be castable to an integral type.

+++ Random_numbers/src/CGAL/Random.cpp
@@ -37,7 +37,7 @@ Random( )
     // get system's time
     std::time_t s;
     std::time( &s);
-    seed = s;
+    seed = (unsigned int)s;
 
     // initialize random numbers generator
     std::srand( seed);
